requirements.txt
fastapi>=0.100.0
gradio>=5.31.0
gradio_imageslider==0.0.17
gradio_client

transformers>=4.37.0,<4.46.0
accelerate>=0.25.0,<0.35.0
tokenizers>=0.15.0,<0.20.0
pydantic>=2.0.0,<3.0.0
Markdown>=3.4.1,<3.7.0
numpy>=1.24.2,<1.27.0
requests>=2.28.2,<2.33.0
sentencepiece==0.1.98
torch>=2.1.0,<2.5.0
torchvision>=0.16.0,<0.20.0
uvicorn>=0.21.1,<0.32.0
wandb>=0.14.0,<0.18.0
httpx>=0.24.0,<0.28.0
scikit-learn>=1.2.2,<1.6.0
einops>=0.7.0,<0.9.0
einops-exts>=0.0.4,<0.1.0
timm>=0.9.8,<1.1.0
openai-clip==1.0.1
fsspec>=2023.4.0,<2024.10.0
kornia>=0.6.9,<0.8.0
matplotlib>=3.7.1,<3.10.0
ninja>=1.11.1,<1.12.0
omegaconf>=2.3.0,<2.4.0
open-clip-torch>=2.26.1,<2.30.0
opencv-python>=4.7.0.72,<4.11.0
pandas>=2.0.1,<2.3.0
Pillow>=9.4.0,<11.0.0
pytorch-lightning>=2.1.2,<2.5.0
PyYAML>=6.0,<6.1.0
scipy>=1.9.1,<1.15.0
tqdm>=4.65.0,<4.67.0
triton>=2.1.0,<3.1.0
urllib3>=1.26.15,<2.3.0
webdataset>=0.2.48,<0.3.0
xformers>=0.0.20,<0.0.30
facexlib>=0.3.0,<0.4.0
k-diffusion>=0.1.1.post1,<0.2.0
diffusers>=0.16.1,<0.32.0